Split pp-ga.ts into core.ts, events.ts, utils.ts
