+++ src/pp-ga/utils.ts
@@ -1,0 +1,13 @@
+import { annotationPrioritiesLabels } from '../api/annotations';
+
+export function formatPriority(priority) {
+  return `${priority} - ${annotationPrioritiesLabels[priority]}`;
+}
+
+export function formatBoolean(val: boolean) {
+  return val ? 'True' : 'False';
+}
+
+export function formatReason(reason) {
+  return `${reason}`;
+}
